+++ projects/angular-sdk/CHANGELOG.md
@@ -5,6 +5,7 @@ Exercise flow:
 - Improved the payment flow UI
 - Restricted the country to Canada and currency to CAD in payment widget
 - Added a warning message regarding the transfer method via bank account limited to Canada
+- Added new parameter: contactFormUrl
 
 Offer dialog:
 - Improved accessibility